pyproject.toml
[tool.poetry]
name = "contact-form"
version = "0.1.0"
description = ""
authors = ["k4ssyi <kasikasi00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.54.1"
uvicorn = "^0.11.3"
python-dotenv = "^0.13.0"
aiofiles = "^0.5.0"
mypy = "^0.770"

[tool.poetry.dev-dependencies]
bpython = "^0.19"
ipdb = "^0.13.2"
isort = "^4.3.21"
black = "^19.10b0"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
